+++ src/X/XRecordOptions.cs
@@ -1,0 +1,11 @@
+namespace Element.X
+{
+    public class XRecordOptions
+    {
+        public string SessionId { get; set; }
+
+        public string MimeType { get; set; } = "audio/webm";
+
+        public int? AudioBitsPerSecond { get; set; }
+    }
+}